[Rename] generator.py 파일 분할하여 두 파일로 나눔
